Refactor ddevapp list stuff into own files
